+++ gostdlib/src/compress/flate.ts
@@ -3,9 +3,9 @@ import type {
   GoInterfaceValue,
 } from "@gotots/runtime/interface-value.js";
 import type { RuntimeSlice } from "@gotots/runtime/slice.js";
-import type { int64, uint8 } from "@gotots/runtime/scalars.js";
+import type { int, uint8 } from "@gotots/gostdlib/internal/scalars.js";
 
 export interface Reader extends GoInterfaceValue {
-  Read(buffer: RuntimeSlice<uint8>): [int64, GoError | undefined];
+  Read(buffer: RuntimeSlice<uint8>): [int, GoError | undefined];
   ReadByte(): [uint8, GoError | undefined];
 }
