+++ gostdlib/src/internal/facets/recovery-io.ts
@@ -1,7 +1,7 @@
 import type { GoError } from "@gotots/runtime/interface-value.js";
 import type { GoRecovery } from "@gotots/runtime/panic.js";
 import type { RuntimeSlice } from "@gotots/runtime/slice.js";
-import type { int64, uint8 } from "@gotots/runtime/scalars.js";
+import type { int, uint8 } from "@gotots/gostdlib/internal/scalars.js";
 
 import { Reader as BufioReader, Writer as BufioWriter } from "../../bufio.js";
 import { Buffer as BytesBuffer } from "../../bytes.js";
@@ -14,7 +14,7 @@ export function BufioReaderRead(
   receiver: BufioReader | undefined,
   destination: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return BufioReader.Read(receiver, destination);
 }
 
@@ -22,7 +22,7 @@ export function BufioWriterWrite(
   receiver: BufioWriter | undefined,
   source: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return BufioWriter.Write(receiver, source);
 }
 
@@ -30,7 +30,7 @@ export function BytesBufferRead(
   receiver: BytesBuffer | undefined,
   destination: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return BytesBuffer.Read(receiver, destination);
 }
 
@@ -38,7 +38,7 @@ export function BytesBufferWrite(
   receiver: BytesBuffer | undefined,
   source: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return BytesBuffer.Write(receiver, source);
 }
 
@@ -53,7 +53,7 @@ export function GzipReaderRead(
   receiver: GzipReader | undefined,
   destination: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return GzipReader.Read(receiver, destination);
 }
 
@@ -75,7 +75,7 @@ export function OsFileRead(
   receiver: OsFile | undefined,
   destination: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return OsFile.Read(receiver, destination);
 }
 
@@ -83,7 +83,7 @@ export function OsFileWrite(
   receiver: OsFile | undefined,
   source: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return OsFile.Write(receiver, source);
 }
 
@@ -91,7 +91,7 @@ export function StringsBuilderWrite(
   receiver: StringsBuilder | undefined,
   source: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return StringsBuilder.Write(receiver, source);
 }
 
@@ -99,6 +99,6 @@ export function StringsReaderRead(
   receiver: StringsReader | undefined,
   destination: RuntimeSlice<uint8>,
   _recovery?: GoRecovery,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   return StringsReader.Read(receiver, destination);
 }
