+++ gostdlib/src/io.ts
@@ -1,7 +1,7 @@
 import type { GoError, GoInterfaceValue } from "@gotots/runtime/interface-value.js";
 import { GoPanic } from "@gotots/runtime/panic.js";
 import { RuntimeSlice } from "@gotots/runtime/slice.js";
-import type { int64, uint8 } from "@gotots/runtime/scalars.js";
+import type { int, uint8 } from "@gotots/gostdlib/internal/scalars.js";
 
 import { New } from "./errors.js";
 import {
@@ -18,11 +18,11 @@ export interface Closer extends GoInterfaceValue {
 }
 
 export interface Reader extends GoInterfaceValue {
-  Read(buffer: RuntimeSlice<uint8>): [int64, GoError | undefined];
+  Read(buffer: RuntimeSlice<uint8>): [int, GoError | undefined];
 }
 
 export interface Writer extends GoInterfaceValue {
-  Write(buffer: RuntimeSlice<uint8>): [int64, GoError | undefined];
+  Write(buffer: RuntimeSlice<uint8>): [int, GoError | undefined];
 }
 
 export interface ReadCloser extends Reader, Closer {}
@@ -40,8 +40,8 @@ class DiscardWriter extends ProviderInterfaceValue implements Writer {
     super(discardType);
   }
 
-  Write(buffer: RuntimeSlice<uint8>): [int64, GoError | undefined] {
-    return [buffer.length, undefined];
+  Write(buffer: RuntimeSlice<uint8>): [int, GoError | undefined] {
+    return [BigInt(buffer.length), undefined];
   }
 }
 
@@ -64,7 +64,7 @@ export const state: {
 export function ReadFull(
   reader: Reader | undefined,
   buffer: RuntimeSlice<uint8>,
-): [int64, GoError | undefined] {
+): [int, GoError | undefined] {
   if (reader === undefined) {
     GoPanic.raiseRuntime("invalid memory address or nil pointer dereference");
   }
