+++ gostdlib/src/net.ts
@@ -3,7 +3,7 @@ import type {
   GoInterfaceValue,
 } from "@gotots/runtime/interface-value.js";
 import type { RuntimeSlice } from "@gotots/runtime/slice.js";
-import type { gostring, int64, uint8 } from "@gotots/runtime/scalars.js";
+import type { gostring, int, uint8 } from "@gotots/gostdlib/internal/scalars.js";
 
 import type { Time } from "./time.js";
 import { ProviderError } from "./internal/runtime/error.js";
@@ -16,12 +16,12 @@ export interface Addr extends GoInterfaceValue {
 export interface Conn extends GoInterfaceValue {
   Close(): GoError | undefined;
   LocalAddr(): Addr | undefined;
-  Read(b: RuntimeSlice<uint8>): [int64, GoError | undefined];
+  Read(b: RuntimeSlice<uint8>): [int, GoError | undefined];
   RemoteAddr(): Addr | undefined;
   SetDeadline(t: Time): GoError | undefined;
   SetReadDeadline(t: Time): GoError | undefined;
   SetWriteDeadline(t: Time): GoError | undefined;
-  Write(b: RuntimeSlice<uint8>): [int64, GoError | undefined];
+  Write(b: RuntimeSlice<uint8>): [int, GoError | undefined];
 }
 
 export interface Listener extends GoInterfaceValue {
